MULTIARCH-4696: deprovision issues and cleanup prior are resolved (#58979)
* MULTIARCH-4696: refactor the deprovision routine and fix security-group-delete


* MULTIARCH-4696: refactor the deprovision


---------

+++ ci-operator/step-registry/upi/deprovision/powervs/cluster/upi-deprovision-powervs-cluster-commands.sh
@@ -99,18 +99,6 @@ function download_terraform_binary() {
     terraform version
 }
 
-# facilitates scp copy from inside the contianer
-function fix_user_permissions() {
-    # Dev Note: scp in a container needs this fix-up
-    if ! whoami &> /dev/null
-    then
-        if [[ -w /etc/passwd ]]
-        then
-            echo "${USER_NAME:-default}:x:$(id -u):0:${USER_NAME:-default} user:${HOME}:/sbin/nologin" >> /etc/passwd
-        fi
-    fi
-}
-
 # Cleanup prior runs
 # VPC: Load Balancers, images, vm instances
 # PowerVS: images, pvm instances
@@ -168,7 +156,7 @@ function cleanup_prior() {
     echo "Cleaning up the Security Groups"
     ibmcloud is security-groups --vpc "${VPC_NAME}" --resource-group-name "${RESOURCE_GROUP}" --output json \
         | jq -r '[.[] | select(.name | contains("ocp-sec-group"))] | .[]?.name' \
-        | xargs --no-run-if-empty -I {} ibmcloud security-group-delete {} --vpc "${VPC_NAME}" --force\
+        | xargs --no-run-if-empty -I {} ibmcloud is security-group-delete {} --vpc "${VPC_NAME}" --force\
         || true
 
     echo "Cleaning up the VPC Load Balancers"
@@ -202,14 +190,31 @@ function cleanup_prior() {
 # Destroy the cluster based on the set configuration / tfvars
 function destroy_upi_cluster() {
     echo "destroy terraform to build PowerVS UPI cluster"
-    cp "${SHARED_DIR}"/var-multi-arch-upi.tfvars "${IBMCLOUD_HOME}"/ocp4-upi-powervs/var-multi-arch-upi.tfvars 
-    echo "UPI TFVARS copied: ${IBMCLOUD_HOME}"/ocp4-upi-powervs/data/var-multi-arch-upi.tfvars
 
     cp "${CLUSTER_PROFILE_DIR}"/ssh-privatekey "${IBMCLOUD_HOME}"/ocp4-upi-powervs/data/id_rsa
     cp "${CLUSTER_PROFILE_DIR}"/ssh-publickey "${IBMCLOUD_HOME}"/ocp4-upi-powervs/data/id_rsa.pub
     chmod 0600 "${IBMCLOUD_HOME}"/ocp4-upi-powervs/data/id_rsa
 
+    # Loads the tfvars if it exists in the shared directory
+    if [ ! -f "${SHARED_DIR}"/var-multi-arch-upi.tfvars ]
+    then
+        echo "No tfvars provided. exiting..."
+        exit 0
+    fi
+
+    cp "${SHARED_DIR}"/var-multi-arch-upi.tfvars "${IBMCLOUD_HOME}"/ocp4-upi-powervs/var-multi-arch-upi.tfvars 
+    echo "UPI TFVARS copied: ${IBMCLOUD_HOME}"/ocp4-upi-powervs/data/var-multi-arch-upi.tfvars
+
+    # Loads the tfstate if it exists in the shared directory
+    if [ ! -f "${SHARED_DIR}"/terraform.tfstate ]
+    then
+        echo "No tfstate file provided"
+        exit 0
+    fi
+
     cp "${SHARED_DIR}"/terraform.tfstate "${IBMCLOUD_HOME}"/ocp4-upi-powervs/data/terraform.tfstate
+
+    # Destroys the current installation for this run
     cd "${IBMCLOUD_HOME}"/ocp-install-dir/ocp4-upi-powervs && \
         "${IBMCLOUD_HOME}"/ocp-install-dir/terraform init && \
         "${IBMCLOUD_HOME}"/ocp-install-dir/terraform destroy -auto-approve \
@@ -229,13 +234,8 @@ setup_ibmcloud_cli
 download_terraform_binary
 download_automation_code
 login_ibmcloud
-
-if [ -f "${SHARED_DIR}"/var-multi-arch-upi.tfvars ]
-then
-    cleanup_prior
-    fix_user_permissions
-    destroy_upi_cluster
-fi
+cleanup_prior
+destroy_upi_cluster
 
 echo "IBM Cloud PowerVS resources destroyed successfully $(date)"
